requirements.txt
# Web framework
fastapi==0.110.1
uvicorn[standard]==0.29.0

# Azure SDKs
azure-search-documents==11.4.0b7
azure-storage-blob==12.19.1
azure-core==1.30.1

# OpenAI (Azure)
openai==1.23.2

# PDF text extraction
PyPDF2==3.0.1

# Environment loading
python-dotenv==1.0.1

# Optional: better logging, CORS, etc.
httpx==0.27.0

# File uploads
python-multipart==0.0.6

# Data validation and settings management
pydantic==1.10.9
 
# Azure OpenAI 
# azure-ai-openai 
# azure-identity 
 
azure-ai-contentsafety==1.0.0
azure-ai-formrecognizer==3.3.3
azure-common==1.1.28
azure-ai-textanalytics==5.3.0
azure-ai-inference 
azure-core 
